requirements.txt
Django == 3.2.3
djangorestframework == 3.12.4
django-filter == 2.4.0
pytest == 6.2.4
pytest-django == 4.3.0
model-bakery == 1.3.1
python-dotenv == 0.17.1
psycopg2-binary == 2.8.6